refactor chart to be more modular
